:truck:  Rename svg logos after last rebase changes
