+++ src/components/Header/Logo/Logo.jsx
@@ -5,7 +5,7 @@ import { Link } from 'gatsby';
 import { hideVisually } from 'polished';
 
 // images
-import LogoSVG from '../../../images/logo-waq-2021-sticker.svg';
+import LogoSVG from '../../../images/logo-waq-21-sticker.svg';
 
 // styles
 import { Img, LogoWrapperStyle } from './Logo.styles';
